refactor: Improve playlist package by separating auxiliaty methods to helpers.go
